move keybinds to separate files
