+++ keybindings/keybindings.go
@@ -1,0 +1,52 @@
+package keybindings
+
+import (
+	"log"
+
+	"github.com/jroimartin/gocui"
+)
+
+func Set(g *gocui.Gui) {
+	if err := g.SetKeybinding("", gocui.KeyCtrlC, gocui.ModNone, quit); err != nil {
+		log.Panicln(err)
+	}
+	if err := g.SetKeybinding("", gocui.KeyCtrlD, gocui.ModNone, quit); err != nil {
+		log.Panicln(err)
+	}
+	if err := g.SetKeybinding("", 'q', gocui.ModNone, quit); err != nil {
+		log.Panicln(err)
+	}
+	if err := g.SetKeybinding("", 'h', gocui.ModNone, help); err != nil {
+		log.Panicln(err)
+	}
+	if err := g.SetKeybinding("", 'j', gocui.ModNone, cursorDown); err != nil {
+		log.Panicln(err)
+	}
+	if err := g.SetKeybinding("", 'k', gocui.ModNone, cursorUp); err != nil {
+		log.Panicln(err)
+	}
+	if err := g.SetKeybinding("", 'a', gocui.ModNone, upvote); err != nil {
+		log.Panicln(err)
+	}
+	if err := g.SetKeybinding("", 'z', gocui.ModNone, downvote); err != nil {
+		log.Panicln(err)
+	}
+	if err := g.SetKeybinding("", gocui.KeyEnter, gocui.ModNone, enter); err != nil {
+		log.Panicln(err)
+	}
+	if err := g.SetKeybinding("", 'l', gocui.ModNone, link); err != nil {
+		log.Panicln(err)
+	}
+	if err := g.SetKeybinding("", 'f', gocui.ModNone, front); err != nil {
+		log.Panicln(err)
+	}
+	if err := g.SetKeybinding("", 'c', gocui.ModNone, comments); err != nil {
+		log.Panicln(err)
+	}
+	if err := g.SetKeybinding("", 'r', gocui.ModNone, refresh); err != nil {
+		log.Panicln(err)
+	}
+	if err := g.SetKeybinding("", 'i', gocui.ModNone, info); err != nil {
+		log.Panicln(err)
+	}
+}
